pyproject.toml
[project]
name = "memo-al"
version = "0.1.0"
description = "Voice-Controlled Local AI Agent — STT → Intent → Tool Execution"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.85.0",
    "docling-core>=2.73.0",
    "langchain-core>=1.2.28",
    "faster-whisper>=1.0.3",
    "groq>=0.9.0",
    "streamlit>=1.32.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]
